backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.79.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.7",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
]
